CommonAncestorForTests => AbstractTestForAllDescriptions (file renaming)
