+++ Magritte/AbstractTestForAllDescriptions.py
